pyproject.toml
[tool.poetry]
name = "authproject"
version = "0.1.0"
description = ""
authors = ["Rohit <RohitSinghEmail@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.2.0"
autopep8 = "^2.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
